+++ spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/thymeleaf/ThymeleafAutoConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2016 the original author or authors.
+ * Copyright 2012-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,21 +16,23 @@
 
 package org.springframework.boot.autoconfigure.thymeleaf;
 
-import java.lang.reflect.Method;
 import java.util.Collection;
+import java.util.LinkedHashMap;
 
+import javax.annotation.PostConstruct;
 import javax.servlet.Servlet;
 
 import com.github.mxab.thymeleaf.extras.dataattribute.dialect.DataAttributeDialect;
 import nz.net.ultraq.thymeleaf.LayoutDialect;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.thymeleaf.dialect.IDialect;
-import org.thymeleaf.extras.conditionalcomments.dialect.ConditionalCommentsDialect;
 import org.thymeleaf.extras.java8time.dialect.Java8TimeDialect;
 import org.thymeleaf.extras.springsecurity4.dialect.SpringSecurityDialect;
-import org.thymeleaf.spring4.SpringTemplateEngine;
-import org.thymeleaf.spring4.resourceresolver.SpringResourceResourceResolver;
-import org.thymeleaf.spring4.templateresolver.SpringResourceTemplateResolver;
-import org.thymeleaf.spring4.view.ThymeleafViewResolver;
+import org.thymeleaf.spring5.SpringTemplateEngine;
+import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
+import org.thymeleaf.spring5.view.ThymeleafViewResolver;
+import org.thymeleaf.templatemode.TemplateMode;
 import org.thymeleaf.templateresolver.ITemplateResolver;
 
 import org.springframework.beans.factory.ObjectProvider;
@@ -38,16 +40,18 @@ import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.boot.autoconfigure.template.TemplateLocation;
 import org.springframework.boot.autoconfigure.web.ConditionalOnEnabledResourceChain;
 import org.springframework.boot.autoconfigure.web.WebMvcAutoConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.Ordered;
 import org.springframework.util.CollectionUtils;
-import org.springframework.util.ReflectionUtils;
+import org.springframework.util.MimeType;
 import org.springframework.web.servlet.resource.ResourceUrlEncodingFilter;
 
 /**
@@ -61,118 +65,103 @@ import org.springframework.web.servlet.resource.ResourceUrlEncodingFilter;
  */
 @Configuration
 @EnableConfigurationProperties(ThymeleafProperties.class)
-@ConditionalOnClass(SpringTemplateEngine.class)
+@ConditionalOnClass(TemplateMode.class)
 @AutoConfigureAfter(WebMvcAutoConfiguration.class)
 public class ThymeleafAutoConfiguration {
 
 	@Configuration
-	@ConditionalOnMissingClass("org.thymeleaf.templatemode.TemplateMode")
-	static class Thymeleaf2Configuration {
+	@ConditionalOnMissingBean(name = "defaultTemplateResolver")
+	static class DefaultTemplateResolverConfiguration {
 
-		@Configuration
-		@ConditionalOnMissingBean(name = "defaultTemplateResolver")
-		static class DefaultTemplateResolverConfiguration
-				extends AbstractTemplateResolverConfiguration {
+		private static final Log logger = LogFactory
+				.getLog(DefaultTemplateResolverConfiguration.class);
 
-			DefaultTemplateResolverConfiguration(ThymeleafProperties properties,
-					ApplicationContext applicationContext) {
-				super(properties, applicationContext);
-			}
+		private final ThymeleafProperties properties;
 
-			@Bean
-			public SpringResourceResourceResolver thymeleafResourceResolver() {
-				return new SpringResourceResourceResolver();
-			}
+		private final ApplicationContext applicationContext;
 
+		DefaultTemplateResolverConfiguration(ThymeleafProperties properties,
+				ApplicationContext applicationContext) {
+			this.properties = properties;
+			this.applicationContext = applicationContext;
 		}
 
-		@Configuration
-		@ConditionalOnClass({ Servlet.class })
-		@ConditionalOnWebApplication
-		static class Thymeleaf2ViewResolverConfiguration
-				extends AbstractThymeleafViewResolverConfiguration {
-
-			Thymeleaf2ViewResolverConfiguration(ThymeleafProperties properties,
-					SpringTemplateEngine templateEngine) {
-				super(properties, templateEngine);
+		@PostConstruct
+		public void checkTemplateLocationExists() {
+			boolean checkTemplateLocation = this.properties.isCheckTemplateLocation();
+			if (checkTemplateLocation) {
+				TemplateLocation location = new TemplateLocation(
+						this.properties.getPrefix());
+				if (!location.exists(this.applicationContext)) {
+					logger.warn("Cannot find template location: " + location
+							+ " (please add some templates or check "
+							+ "your Thymeleaf configuration)");
+				}
 			}
-
-			@Override
-			protected void configureTemplateEngine(ThymeleafViewResolver resolver,
-					SpringTemplateEngine templateEngine) {
-				resolver.setTemplateEngine(templateEngine);
-			}
-
 		}
 
-		@Configuration
-		@ConditionalOnClass(ConditionalCommentsDialect.class)
-		static class ThymeleafConditionalCommentsDialectConfiguration {
-
-			@Bean
-			@ConditionalOnMissingBean
-			public ConditionalCommentsDialect conditionalCommentsDialect() {
-				return new ConditionalCommentsDialect();
+		@Bean
+		public SpringResourceTemplateResolver defaultTemplateResolver() {
+			SpringResourceTemplateResolver resolver = new SpringResourceTemplateResolver();
+			resolver.setApplicationContext(this.applicationContext);
+			resolver.setPrefix(this.properties.getPrefix());
+			resolver.setSuffix(this.properties.getSuffix());
+			resolver.setTemplateMode(this.properties.getMode());
+			if (this.properties.getEncoding() != null) {
+				resolver.setCharacterEncoding(this.properties.getEncoding().name());
 			}
-
+			resolver.setCacheable(this.properties.isCache());
+			Integer order = this.properties.getTemplateResolverOrder();
+			if (order != null) {
+				resolver.setOrder(order);
+			}
+			resolver.setCheckExistence(this.properties.isCheckTemplate());
+			return resolver;
 		}
 
 	}
 
 	@Configuration
-	@ConditionalOnClass(name = "org.thymeleaf.templatemode.TemplateMode")
-	static class Thymeleaf3Configuration {
+	@ConditionalOnClass({ Servlet.class })
+	@ConditionalOnWebApplication
+	static class ThymeleafViewResolverConfiguration {
 
-		@Configuration
-		@ConditionalOnMissingBean(name = "defaultTemplateResolver")
-		static class DefaultTemplateResolverConfiguration
-				extends AbstractTemplateResolverConfiguration {
+		private final ThymeleafProperties properties;
 
-			DefaultTemplateResolverConfiguration(ThymeleafProperties properties,
-					ApplicationContext applicationContext) {
-				super(properties, applicationContext);
-			}
+		private final SpringTemplateEngine templateEngine;
 
-			@Bean
-			@Override
-			public SpringResourceTemplateResolver defaultTemplateResolver() {
-				SpringResourceTemplateResolver resolver = super.defaultTemplateResolver();
-				Method setCheckExistence = ReflectionUtils.findMethod(resolver.getClass(),
-						"setCheckExistence", boolean.class);
-				ReflectionUtils.invokeMethod(setCheckExistence, resolver,
-						getProperties().isCheckTemplate());
-				return resolver;
-			}
-
+		ThymeleafViewResolverConfiguration(ThymeleafProperties properties,
+				SpringTemplateEngine templateEngine) {
+			this.properties = properties;
+			this.templateEngine = templateEngine;
 		}
 
-		@Configuration
-		@ConditionalOnClass({ Servlet.class })
-		@ConditionalOnWebApplication
-		static class Thymeleaf3ViewResolverConfiguration
-				extends AbstractThymeleafViewResolverConfiguration {
+		@Bean
+		@ConditionalOnMissingBean(name = "thymeleafViewResolver")
+		@ConditionalOnProperty(name = "spring.thymeleaf.enabled", matchIfMissing = true)
+		public ThymeleafViewResolver thymeleafViewResolver() {
+			ThymeleafViewResolver resolver = new ThymeleafViewResolver();
+			resolver.setTemplateEngine(this.templateEngine);
+			resolver.setCharacterEncoding(this.properties.getEncoding().name());
+			resolver.setContentType(appendCharset(this.properties.getContentType(),
+					resolver.getCharacterEncoding()));
+			resolver.setExcludedViewNames(this.properties.getExcludedViewNames());
+			resolver.setViewNames(this.properties.getViewNames());
+			// This resolver acts as a fallback resolver (e.g. like a
+			// InternalResourceViewResolver) so it needs to have low precedence
+			resolver.setOrder(Ordered.LOWEST_PRECEDENCE - 5);
+			resolver.setCache(this.properties.isCache());
+			return resolver;
+		}
 
-			Thymeleaf3ViewResolverConfiguration(ThymeleafProperties properties,
-					SpringTemplateEngine templateEngine) {
-				super(properties, templateEngine);
+		private String appendCharset(MimeType type, String charset) {
+			if (type.getCharset() != null) {
+				return type.toString();
 			}
-
-			@Override
-			protected void configureTemplateEngine(ThymeleafViewResolver resolver,
-					SpringTemplateEngine templateEngine) {
-				Method setTemplateEngine;
-				try {
-					setTemplateEngine = ReflectionUtils.findMethod(resolver.getClass(),
-							"setTemplateEngine",
-							Class.forName("org.thymeleaf.ITemplateEngine", true,
-									resolver.getClass().getClassLoader()));
-				}
-				catch (ClassNotFoundException ex) {
-					throw new IllegalStateException(ex);
-				}
-				ReflectionUtils.invokeMethod(setTemplateEngine, resolver, templateEngine);
-			}
-
+			LinkedHashMap<String, String> parameters = new LinkedHashMap<String, String>();
+			parameters.put("charset", charset);
+			parameters.putAll(type.getParameters());
+			return new MimeType(type, parameters).toString();
 		}
 
 	}
