+++ spring-boot-dependencies/pom.xml
@@ -167,7 +167,7 @@
 		<spring-security-jwt.version>1.0.7.RELEASE</spring-security-jwt.version>
 		<spring-security-oauth.version>2.0.12.RELEASE</spring-security-oauth.version>
 		<spring-session.version>2.0.0.BUILD-SNAPSHOT</spring-session.version>
-		<spring-social.version>2.0.0.M1</spring-social.version>
+		<spring-social.version>2.0.0.M2</spring-social.version>
 		<spring-social-facebook.version>3.0.0.M1</spring-social-facebook.version>
 		<spring-social-linkedin.version>2.0.0.M1</spring-social-linkedin.version>
 		<spring-social-twitter.version>2.0.0.M1</spring-social-twitter.version>
@@ -175,12 +175,12 @@
 		<sqlite-jdbc.version>3.15.1</sqlite-jdbc.version>
 		<statsd-client.version>3.1.0</statsd-client.version>
 		<sun-mail.version>${javax-mail.version}</sun-mail.version>
-		<thymeleaf.version>2.1.5.RELEASE</thymeleaf.version>
-		<thymeleaf-extras-springsecurity4.version>2.1.3.RELEASE</thymeleaf-extras-springsecurity4.version>
-		<thymeleaf-extras-conditionalcomments.version>2.1.2.RELEASE</thymeleaf-extras-conditionalcomments.version>
-		<thymeleaf-layout-dialect.version>1.4.0</thymeleaf-layout-dialect.version>
-		<thymeleaf-extras-data-attribute.version>1.3</thymeleaf-extras-data-attribute.version>
-		<thymeleaf-extras-java8time.version>2.1.0.RELEASE</thymeleaf-extras-java8time.version>
+		<thymeleaf.version>3.0.3.RELEASE</thymeleaf.version>
+		<thymeleaf-spring5.version>3.0.3.M1</thymeleaf-spring5.version>
+		<thymeleaf-extras-springsecurity4.version>3.0.1.RELEASE</thymeleaf-extras-springsecurity4.version>
+		<thymeleaf-layout-dialect.version>2.1.2</thymeleaf-layout-dialect.version>
+		<thymeleaf-extras-data-attribute.version>2.0.1</thymeleaf-extras-data-attribute.version>
+		<thymeleaf-extras-java8time.version>3.0.0.RELEASE</thymeleaf-extras-java8time.version>
 		<tomcat.version>8.5.11</tomcat.version>
 		<undertow.version>1.4.8.Final</undertow.version>
 		<unboundid-ldapsdk.version>3.2.0</unboundid-ldapsdk.version>
@@ -2182,6 +2182,17 @@
 			</dependency>
 			<dependency>
 				<groupId>org.springframework.social</groupId>
+				<artifactId>spring-social-web-thymeleaf3</artifactId>
+				<version>${spring-social.version}</version>
+				<exclusions>
+					<exclusion>
+						<artifactId>thymeleaf-spring4</artifactId>
+						<groupId>org.thymeleaf</groupId>
+					</exclusion>
+				</exclusions>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.social</groupId>
 				<artifactId>spring-social-facebook</artifactId>
 				<version>${spring-social-facebook.version}</version>
 			</dependency>
@@ -2256,13 +2267,8 @@
 			</dependency>
 			<dependency>
 				<groupId>org.thymeleaf</groupId>
-				<artifactId>thymeleaf-spring4</artifactId>
-				<version>${thymeleaf.version}</version>
-			</dependency>
-			<dependency>
-				<groupId>org.thymeleaf.extras</groupId>
-				<artifactId>thymeleaf-extras-conditionalcomments</artifactId>
-				<version>${thymeleaf-extras-conditionalcomments.version}</version>
+				<artifactId>thymeleaf-spring5</artifactId>
+				<version>${thymeleaf-spring5.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>org.thymeleaf.extras</groupId>
