--- spring-boot-samples/spring-boot-sample-web-thymeleaf3/src/main/java/sample/web/thymeleaf3/InMemoryMessageRepository.java
@@ -1,58 +1,0 @@
-/*
- * Copyright 2012-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sample.web.thymeleaf3;
-
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.atomic.AtomicLong;
-
-/**
- * @author Dave Syer
- */
-public class InMemoryMessageRepository implements MessageRepository {
-
-	private static AtomicLong counter = new AtomicLong();
-
-	private final ConcurrentMap<Long, Message> messages = new ConcurrentHashMap<Long, Message>();
-
-	@Override
-	public Iterable<Message> findAll() {
-		return this.messages.values();
-	}
-
-	@Override
-	public Message save(Message message) {
-		Long id = message.getId();
-		if (id == null) {
-			id = counter.incrementAndGet();
-			message.setId(id);
-		}
-		this.messages.put(id, message);
-		return message;
-	}
-
-	@Override
-	public Message findMessage(Long id) {
-		return this.messages.get(id);
-	}
-
-	@Override
-	public void deleteMessage(Long id) {
-		this.messages.remove(id);
-	}
-
-}
