--- spring-boot-samples/spring-boot-sample-web-thymeleaf3/src/main/java/sample/web/thymeleaf3/Message.java
@@ -1,69 +1,0 @@
-/*
- * Copyright 2012-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sample.web.thymeleaf3;
-
-import java.util.Calendar;
-
-import org.hibernate.validator.constraints.NotEmpty;
-
-/**
- * @author Rob Winch
- */
-public class Message {
-
-	private Long id;
-
-	@NotEmpty(message = "Message is required.")
-	private String text;
-
-	@NotEmpty(message = "Summary is required.")
-	private String summary;
-
-	private Calendar created = Calendar.getInstance();
-
-	public Long getId() {
-		return this.id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public Calendar getCreated() {
-		return this.created;
-	}
-
-	public void setCreated(Calendar created) {
-		this.created = created;
-	}
-
-	public String getText() {
-		return this.text;
-	}
-
-	public void setText(String text) {
-		this.text = text;
-	}
-
-	public String getSummary() {
-		return this.summary;
-	}
-
-	public void setSummary(String summary) {
-		this.summary = summary;
-	}
-}
