--- spring-boot-samples/spring-boot-sample-web-thymeleaf3/src/main/java/sample/web/thymeleaf3/mvc/MessageController.java
@@ -1,92 +1,0 @@
-/*
- * Copyright 2012-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sample.web.thymeleaf3.mvc;
-
-import javax.validation.Valid;
-
-import sample.web.thymeleaf3.Message;
-import sample.web.thymeleaf3.MessageRepository;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-
-/**
- * @author Rob Winch
- * @author Doo-Hwan Kwak
- */
-@Controller
-@RequestMapping("/")
-public class MessageController {
-
-	private final MessageRepository messageRepository;
-
-	public MessageController(MessageRepository messageRepository) {
-		this.messageRepository = messageRepository;
-	}
-
-	@GetMapping
-	public ModelAndView list() {
-		Iterable<Message> messages = this.messageRepository.findAll();
-		return new ModelAndView("messages/list", "messages", messages);
-	}
-
-	@GetMapping("{id}")
-	public ModelAndView view(@PathVariable("id") Message message) {
-		return new ModelAndView("messages/view", "message", message);
-	}
-
-	@GetMapping(params = "form")
-	public String createForm(@ModelAttribute Message message) {
-		return "messages/form";
-	}
-
-	@PostMapping
-	public ModelAndView create(@Valid Message message, BindingResult result,
-			RedirectAttributes redirect) {
-		if (result.hasErrors()) {
-			return new ModelAndView("messages/form", "formErrors", result.getAllErrors());
-		}
-		message = this.messageRepository.save(message);
-		redirect.addFlashAttribute("globalMessage", "Successfully created a new message");
-		return new ModelAndView("redirect:/{message.id}", "message.id", message.getId());
-	}
-
-	@RequestMapping("foo")
-	public String foo() {
-		throw new RuntimeException("Expected exception in controller");
-	}
-
-	@GetMapping(value = "delete/{id}")
-	public ModelAndView delete(@PathVariable("id") Long id) {
-		this.messageRepository.deleteMessage(id);
-		Iterable<Message> messages = this.messageRepository.findAll();
-		return new ModelAndView("messages/list", "messages", messages);
-	}
-
-	@GetMapping(value = "modify/{id}")
-	public ModelAndView modifyForm(@PathVariable("id") Message message) {
-		return new ModelAndView("messages/form", "message", message);
-	}
-
-}
