--- spring-boot-samples/spring-boot-sample-web-thymeleaf3/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-# Allow Thymeleaf templates to be reloaded at dev time
-spring.thymeleaf.cache: false
-spring.thymeleaf.mode: html
