--- spring-boot-samples/spring-boot-sample-web-thymeleaf3/src/main/resources/templates/layout.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-	xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
-<head>
-<title>Layout</title>
-<link rel="stylesheet" th:href="@{/css/bootstrap.min.css}"
-	href="../../css/bootstrap.min.css" />
-</head>
-<body>
-	<div class="container">
-		<div class="navbar">
-			<div class="navbar-inner">
-				<a class="brand"
-					href="https://github.com/ultraq/thymeleaf-layout-dialect">
-					Thymeleaf - Layout </a>
-				<ul class="nav">
-					<li><a th:href="@{/}" href="messages.html"> Messages </a></li>
-				</ul>
-			</div>
-		</div>
-		<h1 layout:fragment="header">Layout</h1>
-		<div layout:fragment="content">Fake content</div>
-	</div>
-</body>
-</html>
