--- spring-boot-samples/spring-boot-sample-web-thymeleaf3/src/main/resources/templates/messages/list.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-	xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-	layout:decorate="layout">
-<head>
-<title>Messages : View all</title>
-</head>
-<body>
-	<h1 layout:fragment="header">Messages : View all</h1>
-	<div layout:fragment="content" class="container">
-		<div class="pull-right">
-			<a href="form.html" th:href="@{/(form)}">Create Message</a>
-		</div>
-		<table class="table table-bordered table-striped">
-			<thead>
-				<tr>
-					<td>ID</td>
-					<td>Created</td>
-					<td>Summary</td>
-				</tr>
-			</thead>
-			<tbody>
-				<tr th:if="${messages.empty}">
-					<td colspan="3">No messages</td>
-				</tr>
-				<tr th:each="message : ${messages}">
-					<td th:text="${message.id}">1</td>
-					<td th:text="${#calendars.format(message.created)}">July 11,
-						2012 2:17:16 PM CDT</td>
-					<td><a href="view.html" th:href="@{'/' + ${message.id}}"
-						th:text="${message.summary}"> The summary </a></td>
-				</tr>
-			</tbody>
-		</table>
-	</div>
-</body>
-</html>
