--- spring-boot-samples/spring-boot-sample-web-thymeleaf3/src/main/resources/templates/messages/view.html
@@ -1,35 +1,0 @@
-<html xmlns:th="http://www.thymeleaf.org"
-	xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-	layout:decorate="layout">
-<head>
-<title>Messages : View</title>
-</head>
-<body>
-	<h1 layout:fragment="header">Messages : Create</h1>
-	<div layout:fragment="content" class="container">
-		<div class="alert alert-success" th:if="${globalMessage}"
-			th:text="${globalMessage}">Some Success message</div>
-		<div class="pull-right">
-			<a th:href="@{/}" href="list.html"> Messages </a>
-		</div>
-		<dl>
-			<dt>ID</dt>
-			<dd id="id" th:text="${message.id}">123</dd>
-			<dt>Date</dt>
-			<dd id="created" th:text="${#calendars.format(message.created)}">
-				July 11, 2012 2:17:16 PM CDT</dd>
-			<dt>Summary</dt>
-			<dd id="summary" th:text="${message.summary}">A short summary...
-			</dd>
-			<dt>Message</dt>
-			<dd id="text" th:text="${message.text}">A detailed message that
-				is longer than the summary.</dd>
-		</dl>
-		<div class="pull-left">
-			<a href="messages" th:href="@{'/delete/' + ${message.id}}">
-				delete </a> | <a href="form.html"
-				th:href="@{'/modify/' + ${message.id}}"> modify </a>
-		</div>
-	</div>
-</body>
-</html>
