+++ spring-boot-starters/spring-boot-starter-social-twitter/pom.xml
@@ -40,6 +40,16 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.social</groupId>
+			<artifactId>spring-social-web-thymeleaf3</artifactId>
+			<exclusions>
+				<exclusion>
+					<artifactId>thymeleaf</artifactId>
+					<groupId>org.thymeleaf</groupId>
+				</exclusion>
+			</exclusions>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.social</groupId>
 			<artifactId>spring-social-twitter</artifactId>
 		</dependency>
 	</dependencies>
